refactor PickupList
