+++ src/components/Pickups/PickupList.vue
@@ -1,14 +1,20 @@
 <template>
   <div>
-    <PickupItem
-      v-for="pickup in pickups"
-      :key="pickup.id"
-      :pickup="pickup"
-      @join="$emit('join', arguments[0])"
-      @leave="$emit('leave', arguments[0])"
-    >
-      {{ $d(pickup.date, 'dateLongWithDayName') }}
-    </PickupItem>
+    <transition name="slide-toggle">
+      <transition-group
+        name="list"
+        tag="div"
+      >
+        <PickupItem
+          v-for="pickup in pickups"
+          :key="pickup.id"
+          :pickup="pickup"
+          :store-link="storeLink"
+          @join="$emit('join', arguments[0])"
+          @leave="$emit('leave', arguments[0])"
+        />
+      </transition-group>
+    </transition>
   </div>
 </template>
 
@@ -21,6 +27,10 @@ export default {
       type: Array,
       required: true,
     },
+    storeLink: {
+      type: Boolean,
+      default: false,
+    },
   },
   components: {
     PickupItem,
@@ -29,4 +39,10 @@ export default {
 </script>
 
 <style scoped lang="stylus">
+@import '~slidetoggle'
+.list-leave-active
+  transition: all .5s
+.list-leave-to
+  opacity: 0
+  transform: translateX(-50px)
 </style>
