+++ src/pages/Group/Pickups.vue
@@ -1,18 +1,11 @@
 <template>
   <div>
-    <PickupItem
-      v-for="pickup in pickups"
-      :key="pickup.id"
-      :pickup="pickup"
+    <PickupList
+      :pickups="pickups"
+      store-link
       @join="join"
       @leave="leave"
-    >
-      <strong v-if="pickup.store">
-        <router-link :to="{ name: 'store', params: { storeId: pickup.store.id }}">
-          {{ pickup.store.name }}
-        </router-link>
-      </strong> {{ $d(pickup.date, 'dateWithDayName') }}
-    </PickupItem>
+    />
     <KNotice v-if="!hasPickups" >
       <template slot="icon">
         <i class="fas fa-bed"/>
@@ -35,7 +28,7 @@
 </template>
 
 <script>
-import PickupItem from '@/components/Pickups/PickupItem'
+import PickupList from '@/components/Pickups/PickupList'
 import KNotice from '@/components/General/KNotice'
 import StoreList from '@/components/Store/StoreList'
 import { QCard, QCardTitle, QCardMain } from 'quasar'
@@ -46,7 +39,7 @@ import {
 } from 'vuex'
 
 export default {
-  components: { QCard, QCardTitle, QCardMain, PickupItem, KNotice, StoreList },
+  components: { QCard, QCardTitle, QCardMain, PickupList, KNotice, StoreList },
   methods: {
     ...mapActions({
       join: 'pickups/join',
